+++ hutool-core/src/test/java/cn/hutool/core/bean/BeanDescTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 
 /**
  * {@link BeanDesc} 单元测试类
- * 
+ *
  * @author looly
  *
  */
@@ -20,48 +20,48 @@ public class BeanDescTest {
 		Assert.assertEquals("getAge", desc.getGetter("age").getName());
 		Assert.assertEquals("setAge", desc.getSetter("age").getName());
 		Assert.assertEquals(1, desc.getSetter("age").getParameterTypes().length);
-		Assert.assertEquals(int.class, desc.getSetter("age").getParameterTypes()[0]);
-		
+		Assert.assertSame(int.class, desc.getSetter("age").getParameterTypes()[0]);
+
 	}
-	
+
 	@Test
 	public void propDescTes2() {
 		BeanDesc desc = BeanUtil.getBeanDesc(User.class);
-		
+
 		PropDesc prop = desc.getProp("name");
 		Assert.assertEquals("name", prop.getFieldName());
 		Assert.assertEquals("getName", prop.getGetter().getName());
 		Assert.assertEquals("setName", prop.getSetter().getName());
 		Assert.assertEquals(1, prop.getSetter().getParameterTypes().length);
-		Assert.assertEquals(String.class, prop.getSetter().getParameterTypes()[0]);
+		Assert.assertSame(String.class, prop.getSetter().getParameterTypes()[0]);
 	}
-	
+
 	@Test
 	public void propDescOfBooleanTest() {
 		BeanDesc desc = BeanUtil.getBeanDesc(User.class);
-		
+
 		Assert.assertEquals("isAdmin", desc.getGetter("isAdmin").getName());
 		Assert.assertEquals("setAdmin", desc.getSetter("isAdmin").getName());
 		Assert.assertEquals("isGender", desc.getGetter("gender").getName());
 		Assert.assertEquals("setGender", desc.getSetter("gender").getName());
 	}
-	
+
 	@Test
 	public void propDescOfBooleanTest2() {
 		BeanDesc desc = BeanUtil.getBeanDesc(User.class);
-		
+
 		Assert.assertEquals("isIsSuper", desc.getGetter("isSuper").getName());
 		Assert.assertEquals("setIsSuper", desc.getSetter("isSuper").getName());
 	}
-	
+
 	@Test
 	public void getSetTest() {
 		BeanDesc desc = BeanUtil.getBeanDesc(User.class);
-		
+
 		User user = new User();
 		desc.getProp("name").setValue(user, "张三");
 		Assert.assertEquals("张三", user.getName());
-		
+
 		Object value = desc.getProp("name").getValue(user);
 		Assert.assertEquals("张三", value);
 	}
@@ -101,7 +101,7 @@ public class BeanDescTest {
 		public void setAdmin(boolean isAdmin) {
 			this.isAdmin = isAdmin;
 		}
-		
+
 		public boolean isIsSuper() {
 			return isSuper;
 		}